+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+- Dark mode flash on startup: Set GTK dark theme preference and WebView background color before window/content creation
+
 ## [0.12.0] - 2025-10-16
 
 ### Added